requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.103.2
fastavro==1.8.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib-metadata==6.7.0
numpy==1.21.6
pandas==1.3.5
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2025.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
zipp==3.15.0
